pyproject.toml
[tool.poetry]
name = "cds-netsuite-sheet-validator"
version = "0.1.0"
description = ""
authors = ["avickyadav <ashishvickyadav@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-multipart = "^0.0.16"
aiofiles = "^24.1.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
httpx = "^0.27.2"
pyjwt = "^2.9.0"
cryptography = "^43.0.3"
requests = "^2.32.3"
pyodbc = "^5.2.0"
sqlalchemy = "^2.0.36"
urllib3 = "^2.2.3"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
